+++ src/Factory/MessageFactory.php
@@ -10,7 +10,7 @@ use Nyholm\Psr7\Response;
 /**
  * @author Tobias Nyholm <tobias.nyholm@gmail.com>
  */
-class MessageFactory implements \Http\Message\MessageFactory
+final class MessageFactory implements \Http\Message\MessageFactory
 {
     public function createRequest(
         $method,
