+++ src/Request.php
@@ -11,21 +11,12 @@ use Psr\Http\Message\StreamInterface;
 use Psr\Http\Message\UriInterface;
 
 /**
- * @author Michael Dowling and contributors to guzzlehttp/psr7
  * @author Tobias Nyholm <tobias.nyholm@gmail.com>
  */
-class Request implements RequestInterface
+final class Request implements RequestInterface
 {
     use MessageTrait;
-
-    /** @var string */
-    private $method;
-
-    /** @var null|string */
-    private $requestTarget;
-
-    /** @var null|UriInterface */
-    private $uri;
+    use RequestTrait;
 
     /**
      * @param string                               $method  HTTP method
@@ -57,95 +48,5 @@ class Request implements RequestInterface
         if ('' !== $body && null !== $body) {
             $this->stream = (new StreamFactory())->createStream($body);
         }
-    }
-
-    public function getRequestTarget(): string
-    {
-        if (null !== $this->requestTarget) {
-            return $this->requestTarget;
-        }
-
-        $target = $this->uri->getPath();
-        if ('' === $target) {
-            $target = '/';
-        }
-        if ('' !== $this->uri->getQuery()) {
-            $target .= '?'.$this->uri->getQuery();
-        }
-
-        return $target;
-    }
-
-    public function withRequestTarget($requestTarget): self
-    {
-        if (preg_match('#\s#', $requestTarget)) {
-            throw new InvalidArgumentException('Invalid request target provided; cannot contain whitespace');
-        }
-
-        $new = clone $this;
-        $new->requestTarget = $requestTarget;
-
-        return $new;
-    }
-
-    public function getMethod(): string
-    {
-        return $this->method;
-    }
-
-    public function withMethod($method): self
-    {
-        if (!is_string($method)) {
-            throw new \InvalidArgumentException('Method must be a string');
-        }
-
-        $new = clone $this;
-        $new->method = $method;
-
-        return $new;
-    }
-
-    public function getUri(): UriInterface
-    {
-        return $this->uri;
-    }
-
-    public function withUri(UriInterface $uri, $preserveHost = false): self
-    {
-        if ($uri === $this->uri) {
-            return $this;
-        }
-
-        $new = clone $this;
-        $new->uri = $uri;
-
-        if (!$preserveHost || !$this->hasHeader('Host')) {
-            $new->updateHostFromUri();
-        }
-
-        return $new;
-    }
-
-    private function updateHostFromUri(): void
-    {
-        $host = $this->uri->getHost();
-
-        if ('' === $host) {
-            return;
-        }
-
-        if (null !== ($port = $this->uri->getPort())) {
-            $host .= ':'.$port;
-        }
-
-        if (isset($this->headerNames['host'])) {
-            $header = $this->headerNames['host'];
-        } else {
-            $header = 'Host';
-            $this->headerNames['host'] = 'Host';
-        }
-        // Ensure Host is the first header.
-        // See: http://tools.ietf.org/html/rfc7230#section-5.4
-        $this->headers = [$header => [$host]] + $this->headers;
     }
 }
