+++ src/UploadedFile.php
@@ -14,7 +14,7 @@ use RuntimeException;
  * @author Michael Dowling and contributors to guzzlehttp/psr7
  * @author Tobias Nyholm <tobias.nyholm@gmail.com>
  */
-class UploadedFile implements UploadedFileInterface
+final class UploadedFile implements UploadedFileInterface
 {
     /** @var int[] */
     private static $errors = [
